HopsDAO functional
HopsDAO is coded and functional. As I write this commit I realize I probably should have used this opportunity to use JUnit tests to make sure it worked (instead of the by-hand testing I did). I intend to do that going forward. The main function does the same as before on the 'hops' command, though it now uses the DAO instead of the naive statements used originally.